refactor(frontend): extract form field builders + fix verification findings
- Extract constants, field builders, SectionHeader into listing_form_fields.dart
- listing_form.dart reduced from 868 to 527 lines
- Add 4th auto-save test (debounce timing verification)
- Remove dead code in browse_pagination_test.dart

<em>🤖 Co-created with Claude</em>
