+++ frontend/lib/widgets/listing_form.dart
@@ -14,6 +14,7 @@ import '../providers/photo_provider.dart';
 import '../services/api_client.dart';
 import '../utils/logger.dart';
 import '../utils/platform.dart';
+import 'listing_form_fields.dart';
 import 'listing_form_validators.dart';
 import 'photo_section.dart';
 
@@ -57,38 +58,6 @@ class _ListingFormState extends ConsumerState<ListingForm> {
   // ---- Auto-save internals (not in Freezed — mutable/non-serializable) -----
   Timer? _debounce;
   final Map<String, dynamic> _pendingChanges = {};
-
-  // ---- Constants -----------------------------------------------------------
-  static const _cities = ['New York', 'Los Angeles', 'Chicago'];
-  static const _boroughs = [
-    'Manhattan', 'Brooklyn', 'Queens', 'Bronx', 'Staten Island',
-  ];
-  static const _rentalTypes = [
-    ('sublet', 'Sublet'),
-    ('new_lease', 'New Lease'),
-    ('month_to_month', 'Month to Month'),
-    ('short_term', 'Short Term'),
-  ];
-  static const _roomTypes = [
-    ('private_room', 'Private Room'),
-    ('shared_room', 'Shared Room'),
-    ('entire_place', 'Entire Place'),
-  ];
-  static const _veganHouseholds = [
-    ('fully_vegan', 'Fully vegan household'),
-    ('mixed_household', 'Mixed household'),
-  ];
-  static const _furnishedOptions = [
-    ('unfurnished', 'Unfurnished'),
-    ('partially_furnished', 'Partially furnished'),
-    ('fully_furnished', 'Fully furnished'),
-  ];
-  static const _relationships = [
-    ('owner', 'I own the space'),
-    ('manager', 'I manage the space'),
-    ('tenant', 'I am the current tenant'),
-    ('roommate', 'I am a current roommate'),
-  ];
 
   @override
   void initState() {
@@ -429,386 +398,76 @@ class _ListingFormState extends ConsumerState<ListingForm> {
                 ],
               ),
               const SizedBox(height: 24),
-              const _SectionHeader('Basic Information'),
-              _field(
-                label: 'Title *',
-                hint: 'Cozy room in vegan-friendly house',
-                controller: _title,
-                onChanged: (v) => _onChange('title', v),
-              ),
+              const SectionHeader('Basic Information'),
+              buildFormField(label: 'Title *', hint: 'Cozy room in vegan-friendly house', controller: _title, onChanged: (v) => _onChange('title', v)),
               const SizedBox(height: 12),
-              _field(
-                label: 'Description *',
-                hint:
-                    'Describe your space, lifestyle, and what you\'re looking for...',
-                controller: _description,
-                maxLines: 4,
-                onChanged: (v) => _onChange('description', v),
-              ),
+              buildFormField(label: 'Description *', hint: 'Describe your space, lifestyle, and what you\'re looking for...', controller: _description, maxLines: 4, onChanged: (v) => _onChange('description', v)),
               const SizedBox(height: 24),
-              const _SectionHeader('Location'),
-              _dropdown<String>(
-                label: 'City *',
-                value: _fs.city,
-                items: _cities.map((c) => (c, c)).toList(),
-                onChanged: (v) {
-                  setState(() {
-                    _fs = _fs.copyWith(city: v);
-                    if (v != 'New York') _fs = _fs.copyWith(borough: null);
-                  });
-                  _onChange('city', v);
-                  if (v != 'New York') _onChange('borough', null);
-                },
-              ),
+              const SectionHeader('Location'),
+              buildFormDropdown<String>(label: 'City *', value: _fs.city, items: listingCities.map((c) => (c, c)).toList(), onChanged: (v) { setState(() { _fs = _fs.copyWith(city: v); if (v != 'New York') _fs = _fs.copyWith(borough: null); }); _onChange('city', v); if (v != 'New York') _onChange('borough', null); }),
               if (_fs.city == 'New York') ...[
                 const SizedBox(height: 12),
-                _dropdown<String>(
-                  label: 'Borough *',
-                  value: _fs.borough,
-                  items: _boroughs.map((b) => (b, b)).toList(),
-                  onChanged: (v) {
-                    setState(() => _fs = _fs.copyWith(borough: v));
-                    _onChange('borough', v);
-                  },
-                ),
+                buildFormDropdown<String>(label: 'Borough *', value: _fs.borough, items: listingBoroughs.map((b) => (b, b)).toList(), onChanged: (v) { setState(() => _fs = _fs.copyWith(borough: v)); _onChange('borough', v); }),
               ],
               const SizedBox(height: 12),
-              _field(
-                label: 'Neighborhood',
-                hint: 'e.g. Williamsburg, Silver Lake, Wicker Park',
-                controller: _neighborhood,
-                onChanged: (v) => _onChange('neighborhood', v),
-              ),
+              buildFormField(label: 'Neighborhood', hint: 'e.g. Williamsburg, Silver Lake, Wicker Park', controller: _neighborhood, onChanged: (v) => _onChange('neighborhood', v)),
               const SizedBox(height: 24),
-              const _SectionHeader('Rental Details'),
-              _dropdown<String>(
-                label: 'Vegan Household *',
-                value: _fs.veganHousehold,
-                items: _veganHouseholds.toList(),
-                onChanged: (v) {
-                  setState(() => _fs = _fs.copyWith(veganHousehold: v));
-                  _onChange('vegan_household', v);
-                },
-              ),
+              const SectionHeader('Rental Details'),
+              buildFormDropdown<String>(label: 'Vegan Household *', value: _fs.veganHousehold, items: listingVeganHouseholds.toList(), onChanged: (v) { setState(() => _fs = _fs.copyWith(veganHousehold: v)); _onChange('vegan_household', v); }),
               const SizedBox(height: 12),
-              Row(
-                children: [
-                  Expanded(
-                    child: _dropdown<String>(
-                      label: 'Rental Type *',
-                      value: _fs.rentalType,
-                      items: _rentalTypes.toList(),
-                      onChanged: (v) {
-                        setState(() => _fs = _fs.copyWith(rentalType: v));
-                        _onChange('rental_type', v);
-                      },
-                    ),
-                  ),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: _dropdown<String>(
-                      label: 'Room Type *',
-                      value: _fs.roomType,
-                      items: _roomTypes.toList(),
-                      onChanged: (v) {
-                        setState(() => _fs = _fs.copyWith(roomType: v));
-                        _onChange('room_type', v);
-                      },
-                    ),
-                  ),
-                ],
-              ),
+              Row(children: [
+                Expanded(child: buildFormDropdown<String>(label: 'Rental Type *', value: _fs.rentalType, items: listingRentalTypes.toList(), onChanged: (v) { setState(() => _fs = _fs.copyWith(rentalType: v)); _onChange('rental_type', v); })),
+                const SizedBox(width: 12),
+                Expanded(child: buildFormDropdown<String>(label: 'Room Type *', value: _fs.roomType, items: listingRoomTypes.toList(), onChanged: (v) { setState(() => _fs = _fs.copyWith(roomType: v)); _onChange('room_type', v); })),
+              ]),
               const SizedBox(height: 12),
-              Row(
-                children: [
-                  Expanded(
-                    child: _field(
-                      label: 'Monthly Rent (\$) *',
-                      hint: '1200',
-                      controller: _price,
-                      keyboardType: TextInputType.number,
-                      onChanged: (v) => _onChange('price', int.tryParse(v)),
-                    ),
-                  ),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: _dateField(
-                      label: 'Start Date *',
-                      controller: _startDate,
-                      onPicked: (v) => _onChange('start_date', v),
-                    ),
-                  ),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: _dateField(
-                      label: 'End Date (optional)',
-                      controller: _endDate,
-                      onPicked: (v) => _onChange('end_date', v),
-                    ),
-                  ),
-                ],
-              ),
+              Row(children: [
+                Expanded(child: buildFormField(label: 'Monthly Rent (\$) *', hint: '1200', controller: _price, keyboardType: TextInputType.number, onChanged: (v) => _onChange('price', int.tryParse(v)))),
+                const SizedBox(width: 12),
+                Expanded(child: buildFormDateField(label: 'Start Date *', controller: _startDate, onPicked: (v) => _onChange('start_date', v), context: context)),
+                const SizedBox(width: 12),
+                Expanded(child: buildFormDateField(label: 'End Date (optional)', controller: _endDate, onPicked: (v) => _onChange('end_date', v), context: context)),
+              ]),
               const SizedBox(height: 12),
-              _dropdown<String>(
-                label: 'Furnished Status *',
-                value: _fs.furnished,
-                items: _furnishedOptions.toList(),
-                onChanged: (v) {
-                  setState(() => _fs = _fs.copyWith(furnished: v));
-                  _onChange('furnished', v);
-                },
-              ),
+              buildFormDropdown<String>(label: 'Furnished Status *', value: _fs.furnished, items: listingFurnishedOptions.toList(), onChanged: (v) { setState(() => _fs = _fs.copyWith(furnished: v)); _onChange('furnished', v); }),
               const SizedBox(height: 12),
-              Row(
-                children: [
-                  Expanded(
-                    child: _field(
-                      label: 'Size (optional)',
-                      hint: 'e.g. 12x10 feet, 200 sq ft',
-                      controller: _size,
-                      onChanged: (v) => _onChange('size', v),
-                    ),
-                  ),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: _field(
-                      label: 'Transportation (optional)',
-                      hint: 'e.g. Near L train, Bus route 42',
-                      controller: _transportation,
-                      onChanged: (v) => _onChange('transportation', v),
-                    ),
-                  ),
-                ],
-              ),
+              Row(children: [
+                Expanded(child: buildFormField(label: 'Size (optional)', hint: 'e.g. 12x10 feet, 200 sq ft', controller: _size, onChanged: (v) => _onChange('size', v))),
+                const SizedBox(width: 12),
+                Expanded(child: buildFormField(label: 'Transportation (optional)', hint: 'e.g. Near L train, Bus route 42', controller: _transportation, onChanged: (v) => _onChange('transportation', v))),
+              ]),
               const SizedBox(height: 24),
-              const _SectionHeader('About You'),
-              _field(
-                label: 'About you as the lister *',
-                hint:
-                    'Tell potential tenants about yourself, your lifestyle, interests...',
-                controller: _aboutLister,
-                maxLines: 3,
-                onChanged: (v) => _onChange('about_lister', v),
-              ),
+              const SectionHeader('About You'),
+              buildFormField(label: 'About you as the lister *', hint: 'Tell potential tenants about yourself, your lifestyle, interests...', controller: _aboutLister, maxLines: 3, onChanged: (v) => _onChange('about_lister', v)),
               const SizedBox(height: 12),
-              _dropdown<String>(
-                label: 'Your relationship to this space *',
-                value: _fs.listerRelationship,
-                items: _relationships.toList(),
-                onChanged: (v) {
-                  setState(() => _fs = _fs.copyWith(listerRelationship: v));
-                  _onChange('lister_relationship', v);
-                },
-              ),
+              buildFormDropdown<String>(label: 'Your relationship to this space *', value: _fs.listerRelationship, items: listingRelationships.toList(), onChanged: (v) { setState(() => _fs = _fs.copyWith(listerRelationship: v)); _onChange('lister_relationship', v); }),
               const SizedBox(height: 24),
-              const _SectionHeader('Rental Requirements'),
-              _field(
-                label: 'Requirements for potential tenants *',
-                hint: 'What are you looking for in a tenant?',
-                controller: _rentalRequirements,
-                maxLines: 4,
-                onChanged: (v) => _onChange('rental_requirements', v),
-              ),
+              const SectionHeader('Rental Requirements'),
+              buildFormField(label: 'Requirements for potential tenants *', hint: 'What are you looking for in a tenant?', controller: _rentalRequirements, maxLines: 4, onChanged: (v) => _onChange('rental_requirements', v)),
               const SizedBox(height: 12),
-              _field(
-                label: 'Pet policy *',
-                hint: 'e.g. No pets, Cats ok, Dogs with approval',
-                controller: _petPolicy,
-                maxLines: 2,
-                onChanged: (v) => _onChange('pet_policy', v),
-              ),
+              buildFormField(label: 'Pet policy *', hint: 'e.g. No pets, Cats ok, Dogs with approval', controller: _petPolicy, maxLines: 2, onChanged: (v) => _onChange('pet_policy', v)),
               const SizedBox(height: 24),
-              const _SectionHeader('Photos'),
-              PhotoSection(
-                photos: _fs.photos,
-                uploading: _fs.uploadingPhotos,
-                onAdd: _pickAndUploadPhotos,
-                onDelete: _deletePhoto,
-              ),
+              const SectionHeader('Photos'),
+              PhotoSection(photos: _fs.photos, uploading: _fs.uploadingPhotos, onAdd: _pickAndUploadPhotos, onDelete: _deletePhoto),
               const SizedBox(height: 24),
-              const _SectionHeader('Additional Options'),
-              SwitchListTile(
-                title: const Text('Seeking a roommate (not just a tenant)'),
-                value: _fs.seekingRoommate,
-                onChanged: (v) {
-                  setState(() => _fs = _fs.copyWith(seekingRoommate: v));
-                  _onChange('seeking_roommate', v);
-                },
-                contentPadding: EdgeInsets.zero,
-              ),
-              SwitchListTile(
-                title: const Text('Include phone number in listing'),
-                value: _fs.includePhone,
-                onChanged: (v) {
-                  setState(() => _fs = _fs.copyWith(includePhone: v));
-                  _onChange('include_phone', v);
-                },
-                contentPadding: EdgeInsets.zero,
-              ),
+              const SectionHeader('Additional Options'),
+              SwitchListTile(title: const Text('Seeking a roommate (not just a tenant)'), value: _fs.seekingRoommate, onChanged: (v) { setState(() => _fs = _fs.copyWith(seekingRoommate: v)); _onChange('seeking_roommate', v); }, contentPadding: EdgeInsets.zero),
+              SwitchListTile(title: const Text('Include phone number in listing'), value: _fs.includePhone, onChanged: (v) { setState(() => _fs = _fs.copyWith(includePhone: v)); _onChange('include_phone', v); }, contentPadding: EdgeInsets.zero),
               if (_fs.includePhone) ...[
                 const SizedBox(height: 8),
-                _field(
-                  label: 'Phone Number',
-                  hint: '(555) 123-4567',
-                  controller: _phoneNumber,
-                  keyboardType: TextInputType.phone,
-                  onChanged: (v) => _onChange('phone_number', v),
-                ),
+                buildFormField(label: 'Phone Number', hint: '(555) 123-4567', controller: _phoneNumber, keyboardType: TextInputType.phone, onChanged: (v) => _onChange('phone_number', v)),
               ],
               const SizedBox(height: 32),
-              Row(
-                children: [
-                  FilledButton(
-                    onPressed: _previewOrSaveFirst,
-                    child: const Text('Preview listing'),
-                  ),
-                  const SizedBox(width: 12),
-                  OutlinedButton(
-                    onPressed: () => context.go('/dashboard'),
-                    child: const Text('Save & exit'),
-                  ),
-                ],
-              ),
+              Row(children: [
+                FilledButton(onPressed: _previewOrSaveFirst, child: const Text('Preview listing')),
+                const SizedBox(width: 12),
+                OutlinedButton(onPressed: () => context.go('/dashboard'), child: const Text('Save & exit')),
+              ]),
               const SizedBox(height: 24),
             ],
           ),
         ),
       ),
-    );
-  }
-
-  // ---- Small builders ------------------------------------------------------
-
-  Widget _field({
-    required String label,
-    required String hint,
-    required TextEditingController controller,
-    required void Function(String) onChanged,
-    int maxLines = 1,
-    TextInputType? keyboardType,
-  }) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(label,
-            style:
-                const TextStyle(fontSize: 13, fontWeight: FontWeight.w500)),
-        const SizedBox(height: 4),
-        TextField(
-          controller: controller,
-          maxLines: maxLines,
-          keyboardType: keyboardType,
-          decoration: InputDecoration(
-            hintText: hint,
-            isDense: true,
-            border: const OutlineInputBorder(),
-            contentPadding:
-                const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
-          ),
-          onChanged: onChanged,
-        ),
-      ],
-    );
-  }
-
-  Widget _dropdown<T>({
-    required String label,
-    required T? value,
-    required List<(T, String)> items,
-    required void Function(T?) onChanged,
-  }) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(label,
-            style:
-                const TextStyle(fontSize: 13, fontWeight: FontWeight.w500)),
-        const SizedBox(height: 4),
-        DropdownButtonFormField<T>(
-          key: ValueKey(value),
-          initialValue: value,
-          decoration: const InputDecoration(
-            isDense: true,
-            border: OutlineInputBorder(),
-            contentPadding:
-                EdgeInsets.symmetric(horizontal: 12, vertical: 10),
-          ),
-          items: [
-            const DropdownMenuItem(value: null, child: Text('Select...')),
-            ...items.map((pair) =>
-                DropdownMenuItem(value: pair.$1, child: Text(pair.$2))),
-          ],
-          onChanged: onChanged,
-        ),
-      ],
-    );
-  }
-
-  Widget _dateField({
-    required String label,
-    required TextEditingController controller,
-    required void Function(String?) onPicked,
-  }) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(label,
-            style:
-                const TextStyle(fontSize: 13, fontWeight: FontWeight.w500)),
-        const SizedBox(height: 4),
-        TextField(
-          controller: controller,
-          readOnly: true,
-          decoration: const InputDecoration(
-            hintText: 'YYYY-MM-DD',
-            isDense: true,
-            border: OutlineInputBorder(),
-            contentPadding:
-                EdgeInsets.symmetric(horizontal: 12, vertical: 10),
-            suffixIcon: Icon(Icons.calendar_today, size: 16),
-          ),
-          onTap: () async {
-            final now = DateTime.now();
-            final picked = await showDatePicker(
-              context: context,
-              initialDate: now,
-              firstDate: now.subtract(const Duration(days: 30)),
-              lastDate: now.add(const Duration(days: 730)),
-            );
-            if (picked != null) {
-              final formatted =
-                  '${picked.year}-${picked.month.toString().padLeft(2, '0')}-${picked.day.toString().padLeft(2, '0')}';
-              controller.text = formatted;
-              onPicked(formatted);
-            }
-          },
-        ),
-      ],
-    );
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Section header
-// ---------------------------------------------------------------------------
-
-class _SectionHeader extends StatelessWidget {
-  const _SectionHeader(this.title);
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(title,
-            style: Theme.of(context)
-                .textTheme
-                .titleMedium
-                ?.copyWith(fontWeight: FontWeight.bold)),
-        const Divider(),
-        const SizedBox(height: 12),
-      ],
     );
   }
 }
